requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
chardet==3.0.4
configparser==3.5.0
discord.py==1.0.0a0
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
websockets==5.0.1
yarl==1.2.6
